+++ meveo-admin/web/src/main/java/org/meveo/admin/action/catalog/TriggeredEDRTemplateBean.java
@@ -25,9 +25,11 @@ import javax.inject.Named;
 import org.meveo.admin.action.BaseBean;
 import org.meveo.admin.exception.BusinessException;
 import org.meveo.cache.RatingCacheContainerProvider;
+import org.meveo.model.catalog.ChargeTemplate;
 import org.meveo.model.catalog.TriggeredEDRTemplate;
 import org.meveo.service.base.PersistenceService;
 import org.meveo.service.base.local.IPersistenceService;
+import org.meveo.service.catalog.impl.ChargeTemplateServiceAll;
 import org.meveo.service.catalog.impl.TriggeredEDRTemplateService;
 import org.omnifaces.cdi.ViewScoped;
 import org.primefaces.context.RequestContext;
@@ -98,11 +100,24 @@ public class TriggeredEDRTemplateBean extends BaseBean<TriggeredEDRTemplate> {
 		ratingCacheContainerProvider.updateUsageChargeTemplateInCache(entity);
 		return result;
 	}
+	
+	@Inject
+	private ChargeTemplateServiceAll chargeTemplateServiceAll;
 
 	@Override
 	protected void canDelete() {
 		boolean result=true;
-		this.delete();
+		List<ChargeTemplate> chargeTemplates=chargeTemplateServiceAll.findByEDRTemplate(entity);
+		if(chargeTemplates!=null&&chargeTemplates.size()!=0){
+			for(ChargeTemplate chargeTemplate:chargeTemplates){
+				chargeTemplate.getEdrTemplates().remove(entity);
+				chargeTemplateServiceAll.update(chargeTemplate);
+			}
+		}
+		
+		if(result){
+			this.delete();
+		}
 		RequestContext requestContext = RequestContext.getCurrentInstance();
 		requestContext.addCallbackParam("result", result);
 	}
